requirements.txt
numpy>=1.24.0
Pillow>=10.0.0
imageio>=2.31.0
imageio-ffmpeg>=0.4.9
scipy>=1.11.0
noise>=1.2.2
click>=8.1.0